More work on Paramiko engine
@atoonk and @btoonk identified some cases where Paramiko was disconnecting, but we never noticed.  Checking for an exception on the Transport seems to be the answer (since Transport is actually a seperate Thread.)

Also moved our threads out into their own objects to make it easier to isolate.  This also allows us to properly shutdown the log reader when we disconnect.  The only thing we aren't really cleanly shutting down are the forwarded connections (in the event of tunnel termination.)  I'm not sure how much of an issue that is so not digging into it further at this point.